packages/x402-secure/pyproject.toml
[project]
name = "x402-secure"
version = "0.1.1"
description = "Secure x402 client SDK with agentic payments for buyers and sellers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "t54 labs", email = "dev@t54.ai" }]
dependencies = [
  "httpx>=0.25.0",
  "openai>=2.3.0",
  "opentelemetry-api>=1.24.0",
]

[project.optional-dependencies]
signing = [
  "x402>=0.2.0",
  "eth-account>=0.10.0",
  "eth-utils>=2.3.0",
]
otel = [
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp>=1.24.0",
]
examples = [
  "fastapi>=0.104.0",
  "python-dotenv>=1.0.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.0.0",
  "x402>=0.2.0",
  "eth-account>=0.10.0",
  "eth-utils>=2.3.0",
  "opentelemetry-sdk>=1.24.0",
  "opentelemetry-exporter-otlp>=1.24.0",
]
agent = [
  "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/x402_secure_client"]
